+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 # Change Log
 
-## [v4.19.0](https://github.com/plivo/plivo-php/releases/tag/v4.19.0) - 2021-05-28
-- New Send Message interface along with existing interface support.
+## [v4.21.0](https://github.com/plivo/plivo-php/releases/tag/v4.21.0) - 2021-07-13
+- Power pack ID has been included to the response for the [list all messages API](https://www.plivo.com/docs/sms/api/message/list-all-messages/) and the [get message details API](https://www.plivo.com/docs/sms/api/message#retrieve-a-message).
+- Support for filtering messages by Power pack ID has been added to the [list all messages API](https://www.plivo.com/docs/sms/api/message#list-all-messages).
+
+## [v4.20.0](https://github.com/plivo/plivo-php/releases/tag/v4.20.0) - 2021-07-13
+- Add SDK support for MPC APIs and XML.
+
+## [v4.19.2](https://github.com/plivo/plivo-php/releases/tag/v4.19.2) - 2021-07-05
+- **WARNING**: Remove total_count field from meta data for list MDR response
+
+## [v4.19.1](https://github.com/plivo/plivo-php/releases/tag/v4.19.1) - 2021-07-02
+- Adds VoiceNetworkGroup to List/Get Call response.
+
+## [v4.19.0](https://github.com/plivo/plivo-php/releases/tag/v4.19.0) - 2021-06-15
+- Add stir verification param as part of Get CDR and live call APIs
 
 ## [v4.18.3](https://github.com/plivo/plivo-php/releases/tag/v4.18.3) - 2021-05-12
 - Fixing the listMedia function invoke to fetch the media detail on message_uuid
@@ -21,7 +34,7 @@
 ## [v4.17.1](https://github.com/plivo/plivo-php/releases/tag/v4.17.1) - 2021-02-15
 - Fix PHP v8 deprectaion warning for PlivoResponseException
 
-## [v4.17.0](https://github.com/plivo/plivo-php/releases/tag/v4.17.0) - 2020-12-17
+## [v4.17.0](https://github.com/plivo/plivo-php/releases/tag/v4.16.0) - 2020-12-17
 - Add exception to handle destination param - SMS.
 
 ## [v4.16.0](https://github.com/plivo/plivo-php/releases/tag/v4.16.0) - 2020-12-08